requirements.txt
aiohttp
beautifulsoup4
chromadb
coverage
fastapi
helpsk
httpx
langchain
llama-cpp-python
lxml
nest-asyncio
openai
pydantic
pypdf
pytest
requests
ruff
streamlit
tiktoken
uvicorn

# required for transformers.AutoModelForCausalLM.from_pretrained('mosaicml/mpt-7b-instruct' ...)
einops
